+++ app/api/ada/lib/storage/bridge/IStorageBridge.js
@@ -7,9 +7,9 @@ import { Bip44Wallet } from '../models/Bip44Wallet';
 import { PublicDeriver } from '../models/PublicDeriver';
 
 export interface IStorageBridge {
-  getConceptualWalletData(key: number): Promise<ConceptualWalletRow | typeof undefined>;
-  getPublicDeriverData(key: number): Promise<PublicDeriverRow | typeof undefined>;
-  getBip44WrapperData(key: number): Promise<Bip44WrapperRow | typeof undefined>;
+  getConceptualWalletData(key: number): Promise<ConceptualWalletRow | void>;
+  getPublicDeriverData(key: number): Promise<PublicDeriverRow | void>;
+  getBip44WrapperData(key: number): Promise<Bip44WrapperRow | void>;
 
   addPublicDeriverFunctionality(publicDeriver: PublicDeriver): Promise<void>;
   addBip44WalletFunctionality(bip44Wallet: Bip44Wallet): Promise<void>;
